Dijkstra shortest path refactoring

+++ ProblemSets/ProblemSets/ComputerScience/DataTypes/Edge.cs
@@ -15,7 +15,7 @@
 	public struct EdgeEndPoint
 	{
 		public int To;
-		public int Weight;
+		public long Weight;
 
 		public override string ToString()
 		{
